backend/pyproject.toml
[tool.poetry]
name = "messenger"
version = "0.1.0"
description = ""
authors = ["Focu5nik <evgexabu@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.28.0"}
pydantic = {extras = ["email"], version = "^2.6.3"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0.28"}
aiosqlite = "^0.20.0"
pydantic-settings = "^2.2.1"
alembic = "^1.13.1"
pyjwt = {extras = ["crypto"], version = "^2.8.0"}
bcrypt = "^4.1.2"
python-multipart = "^0.0.9"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
black = "^24.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
